+++ lib/iris/tests/unit/common/mixin/test_pintunits.py
@@ -14,7 +14,7 @@ import numpy as np
 import pytest
 
 import iris.common.mixin
-from iris.common.mixin import PintUnit
+from iris.common.units import PintUnit
 
 
 def test_num2date():
@@ -48,8 +48,8 @@ def test_calendar():
     assert str(unit) == "days since 1970-01-01"
 
 
-_UNKNOWN_NAMES = iris.common.mixin.PintUnit._IRIS_EXTRA_CATEGORIES["unknown"]
-_NOUNIT_NAMES = iris.common.mixin.PintUnit._IRIS_EXTRA_CATEGORIES["no_unit"]
+_UNKNOWN_NAMES = iris.common.units.PintUnit._IRIS_EXTRA_CATEGORIES["unknown"]
+_NOUNIT_NAMES = iris.common.units.PintUnit._IRIS_EXTRA_CATEGORIES["no_unit"]
 
 
 class TestFromUnit:
